fix: rename util and platform to resolve import error
util imports numpy, which conflicts with util and platform. Renaming to
utils and platforms.
