+++ python/helios/validation.py
@@ -1,4 +1,4 @@
-from helios.util import find_file, is_real_iterable
+from helios.utils import find_file, is_real_iterable
 
 from collections.abc import Iterable
 from pathlib import Path
