+++ tests/python/test_util.py
@@ -1,4 +1,4 @@
-from helios.util import *
+from helios.utils import *
 
 from pathlib import Path
